requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pypdf==3.17.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
